runtime: add support for "elastic" goroutines for elastic CockroachDB work
CockroachDB has the goal to provide performance isolation via inter-tenant
fairness, and within tenant prioritization. Currently, this is attempted
purely via admission control mechanisms that attempt to reduce queuing in
the goroutine scheduler (queue of runnable goroutines) by queuing outside,
in admission control queues. We have two classes of work: regular and
elastic, where the goal is that regular work should be able to saturate the
CPU. Elastic work is optional, and it is acceptable to not service any
elastic work if CPU utilization is "very high" (say > 95%). However, we
observe delays in scheduling runnable goroutines for regular traffic, that
exceed 20ms, even when the CPU utilization is < 80%. The changes here are a
prototype to attempt to reduce these delays while allowing CPU utilization
due to elastic work to exceed 90%. NB: elastic work could also be
characterized as latency-insensitive, and that is certainly true for our
case -- we use the term elastic because it is also acceptable for the
throughput of this work to be adjusted in order to utilize available
resources.

We assume that we cannot make very extensive changes to the goroutine
scheduler in order to embed knowledge of tenants and fine-grained
priorities. This would be hard to maintain and introduce more complex
data-structures (like heaps) into the scheduler.

The changes here are meant to be work in conjunction with admission control.
After admission is granted, we have a classification on whether the work
is elastic or not. Regular work continues as usual. Elastic work does:
   func() {
     SetGoroutineAsElastic(true)
     defer SetGoroutineAsElastic(false)
     runtime.Gosched()
     ... do work, with calls to runtime.Gosched() at 2ms intervals ...
   }()

The expectation here is that:
- Elastic goroutines are additive to regular goroutines in the runnable
  queues in the goroutine scheduler, i.e., they are not perturbing how the
  regular goroutines are distributed across the runnable queues when there
  were no elastic G's.

- Whenever a runnable G is being picked for a P from one of these queues, a
  regular G will be preferred.

- If an elastic G starts running because there were no regular G's, and
  regular G's arrive immediately after the elastic G starts running, the
  regular G's will not need to wait for the full forcePreemptNS = 10ms.

- An elastic G will not be starved forever, since admission control only
  admits elastic G's when the CPU utilization is < 95%, so there will be
  idle time where an already admitted elastic G will get to run.

Risks:
- Priority inversion caused by elastic goroutines holding mutexes or
  CockroachDB latches.